# Apply go tweaks to go_con