Applies optimizations from #395 to go_con.

Seems to be slower...(tested twice with same results)

Main:
`Go Concurrent | 12.19 ms | 164.52 ms | 1.43 s`

This PR:
```
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 12.449775ms
        Processing time (w/o IO): 14.2278ms
        Processing time (w/o IO): 12.854681ms
        Processing time (w/o IO): 12.754979ms
        Processing time (w/o IO): 12.81938ms
        Processing time (w/o IO): 12.571376ms
        Processing time (w/o IO): 12.623178ms
        Processing time (w/o IO): 12.324273ms
        Processing time (w/o IO): 14.777608ms
        Processing time (w/o IO): 12.512376ms
        Processing time (w/o IO): 12.364274ms
        Processing time (w/o IO): 12.278273ms
        Processing time (w/o IO): 12.388574ms
          Time (mean ± σ):      52.8 ms ±   1.9 ms    [User: 84.4 ms, System: 9.9 ms]
          Range (min … max):    50.2 ms …  56.7 ms    10 runs
         
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 171.469112ms
        Processing time (w/o IO): 171.561813ms
        Processing time (w/o IO): 172.486126ms
          Time (mean ± σ):     315.1 ms ±   0.4 ms    [User: 809.5 ms, System: 35.4 ms]
          Range (min … max):   314.8 ms … 315.4 ms    2 runs
         
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 1.487239318s
        Processing time (w/o IO): 1.488106605s
        Processing time (w/o IO): 1.494551652s
          Time (mean ± σ):      1.980 s ±  0.011 s    [User: 6.412 s, System: 0.103 s]
          Range (min … max):    1.973 s …  1.988 s    2 runs

```