@jinyus 2023-11-01
After switching back to isize:
```
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 11.972857ms
        Processing time (w/o IO): 12.332162ms
        Processing time (w/o IO): 12.044259ms
        Processing time (w/o IO): 15.231ms
        Processing time (w/o IO): 11.685253ms
        Processing time (w/o IO): 11.606352ms
        Processing time (w/o IO): 11.739954ms
        Processing time (w/o IO): 11.741455ms
        Processing time (w/o IO): 11.523651ms
        Processing time (w/o IO): 11.554352ms
        Processing time (w/o IO): 11.805355ms
        Processing time (w/o IO): 11.696154ms
        Processing time (w/o IO): 11.597552ms
          Time (mean ± σ):      51.4 ms ±   1.7 ms    [User: 80.3 ms, System: 10.3 ms]
          Range (min … max):    49.6 ms …  55.4 ms    10 runs
         
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 164.622062ms
        Processing time (w/o IO): 163.027342ms
        Processing time (w/o IO): 163.653349ms
          Time (mean ± σ):     314.8 ms ±   3.4 ms    [User: 792.3 ms, System: 36.2 ms]
          Range (min … max):   312.4 ms … 317.2 ms    2 runs
         
Go Concurrent:

        Benchmark 1: ./related_concurrent
        Processing time (w/o IO): 1.430885231s
        Processing time (w/o IO): 1.429477572s
        Processing time (w/o IO): 1.414835564s
          Time (mean ± σ):      1.911 s ±  0.008 s    [User: 6.125 s, System: 0.104 s]
          Range (min … max):    1.905 s …  1.917 s    2 runs
```